requirements.txt
# Core ML
torch
torchvision
scikit-learn
numpy
pandas
matplotlib

# Backend
fastapi
uvicorn

# Frontend
streamlit
folium
streamlit-folium

# Utilities
opencv-python
Pillow
requests
